requirements.txt
Flask==1.1.1
braintree>=4.0.0
stripe>=2.48.0
MarkupSafe>=0.23
mock>=1.3.0
Werkzeug>=0.7
Jinja2>=2.4
itsdangerous>=0.21
requests<3.0,>=0.11.1
pbr>=0.11
six>=1.7
python-dotenv==0.12.0
gunicorn==20.0.4
pymongo[srv]==3.10.1